Add files via upload
New kernel to calculate Boltzmann factor $p_ik$ using blocks of warps to avoid problems with a maximum size of tracks due to the maximum number of threads per block. Tracks are split into blocks with the size of the warp and the number of blocks is multiplied by the number of the vertex.