Add files via upload
New kernel to calculate the sum of the Boltzmann factor denominator ($\sum_k' rho_k' exp(E_ik')$), the denominator of $p_ik$ using blocks of warps to avoid problems with a maximum size of tracks due to the maximum number of threads per block. The denominator elements are split into blocks with the size of the warp  (the total number of blocks is multiplied by the number of the vertex). After that, a parallel reduction is made summing all elements with the same "i" (tracks index).